moved logger utils
