+++ packages/forgae-cli/commands.js
@@ -21,7 +21,8 @@ const node = require('./forgae-node/node.js');
 const deploy = require('./forgae-deploy/deploy.js');
 const config = require('./config.json');
 const history = require('../forgae-logger/logger-service/log-store-service');
-const printReportTable = require('../forgae-logger/utils').printReportTable;
+const printReportTable = require('forgae-utils').printReportTable;
+// const printReportTable = require('../forgae-utils/logger-utils').printReportTable;
 const contracts = require('./forgae-contracts/forgae-contracts.js');
 const shape = require('./forgae-shapes/shape-commander');
