+++ packages/forgae-cli/package.json
@@ -33,7 +33,9 @@
     "promisify-child-process": "^2.1.2",
     "shelljs": "^0.8.2",
     "simple-git": "^1.107.0",
-    "sinon": "^7.1.1"
+    "sinon": "^7.1.1",
+    "forgae-utils": "1.0.0"
+    
   },
   "devDependencies": {
     "@typescript-eslint/parser": "^1.7.0",
