+++ packages/forgae-logger/package.json
@@ -1,0 +1,12 @@
+{
+  "name": "forgae-logger",
+  "version": "1.0.0",
+  "description": "",
+  "main": "utils.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC"
+}
