+++ packages/forgae-utils/index.js
@@ -1,0 +1,6 @@
+const {printReportTable, getReadableStatus} = require('./logger-utils')
+
+module.exports = {
+    printReportTable,
+    getReadableStatus
+}
